requirements.txt
awscli==1.25.64
botocore==1.27.63
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.4
docutils==0.16
idna==3.3
jmespath==1.0.1
numpy==1.23.2
pandas==1.4.3
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==5.4.1
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
